requirements.txt
requests>=2.28
openai
python-dotenv